Add: 1password to yafti
The reason why is 1password is installed into `/opt` which isnt mounted
at container build time. this means it fails cause it hangs. related:
https://github.com/orgs/ublue-os/discussions/114 . I believe yafti
should fix this because its able to install during run time. I may do
this with mullvad.

yafti clearly isnt designed for this but can be hacked together